Refactor CRUD operations
The app had an infinite loop of calls to the blog API when initializing the list on the frontend because the useEffect depended on it and, at the same time, updated it. The backend has been refactored so that creating and modifying a blog now returns information about the user who created it, making it easier to work with on the frontend.